Complete re-factor
The code was turning into a complete mess and it wasn't clear what was happening at all and definitely not reproducible. I rewrote everything to simply the code and reduce the complexity of the workspace. I focussed on getting a few key model types and showing how they work and how they don't. The new code is much easier to understand and run.